feat(brett): A5 lazy board-boot + Hauptmenü entry screen [T000459]
Move boot() verbatim into board-boot.ts (bootBoard); rewrite main.ts thin and
Three-free, mounting the scene lazily via import('./board-boot') only on first
board-view entry. Add #brett-menu overlay root + guard the legacy join-overlay
auto-mount. Build now splits Three.js: index 558kB -> 16kB, board-boot 551kB
deferred chunk.
